Implement save game mechanism

+++ game/items/gameitems.cpp
@@ -80,7 +80,7 @@ QVariant GameItems::data(const QModelIndex &index, int role) const
     } else if (role == ItemTypeRole) {
         return gameItem->itemType();
     } else if (role == LayerRole) {
-        return static_cast<int>(gameItem->layer());
+        return gameItem->layer();
     } else if (role == ItemIdRole) {
         return gameItem->itemId();
     } else if (role == PriceRole) {
